+++ src/factor_instance_provider/factor_instance_provider.rs
@@ -6,127 +6,6 @@ use rand::seq::index;
 
 use crate::prelude::*;
 
-/// The reason why a request is unfulfillable, and if the reason is that the
-/// last factor would be consumed, the value of that last factor is included,
-/// to act as the range.
-#[derive(Clone, PartialEq, Eq, Hash, Debug, EnumAsInner)]
-pub enum UnfulfillableRequestReason {
-    /// Users before Radix Wallet 2.0 does not have any cache.
-    /// This will be kick of the cumbersome process of analyzing the Profile
-    /// and deriving a broad range of keys to find out the "last used" key per
-    /// factor source, and then use that to derive the next batch of keys and
-    /// cache them.
-    Empty,
-
-    /// The request would consume the last factor, the `HDPathComponent` is
-    /// the value of this last factor, which we can use as a base for the
-    /// next index range to derive keys for, i.e. we will derive keys in the range
-    /// `(last_index + 1, last_index + N)` where `N` is the batch size (e.g. 50).
-    Last(HDPathComponent),
-}
-
-/// A request that cannot be fulfilled, and the reason why.
-#[derive(Clone, PartialEq, Eq, Hash, Debug)]
-pub struct UnfulfillableRequest {
-    /// The request which cannot be fulfilled.
-    request: DerivationRequest,
-
-    /// The reason why `request` could not be fulfilled.
-    reason: UnfulfillableRequestReason,
-}
-impl UnfulfillableRequest {
-    pub fn empty(request: DerivationRequest) -> Self {
-        Self {
-            request,
-            reason: UnfulfillableRequestReason::Empty,
-        }
-    }
-
-    /// # Panics
-    /// Panics if `last_factor` does not share same parameters as `request`
-    pub fn last(
-        request: DerivationRequest,
-        last_factor: &HierarchicalDeterministicFactorInstance,
-    ) -> Self {
-        assert!(
-            last_factor.matches(&request),
-            "last_factor must match request"
-        );
-        Self {
-            request,
-            reason: UnfulfillableRequestReason::Last(last_factor.derivation_path().index),
-        }
-    }
-
-    pub fn is_reason_empty(&self) -> bool {
-        matches!(self.reason, UnfulfillableRequestReason::Empty)
-    }
-
-    pub fn is_reason_last(&self) -> bool {
-        matches!(self.reason, UnfulfillableRequestReason::Last(_))
-    }
-}
-
-impl HierarchicalDeterministicFactorInstance {
-    fn matches(&self, request: &DerivationRequest) -> bool {
-        self.factor_source_id() == request.factor_source_id
-            && self.derivation_path().matches(request)
-    }
-}
-impl DerivationPath {
-    fn matches(&self, request: &DerivationRequest) -> bool {
-        self.network_id == request.network_id
-            && self.entity_kind == request.entity_kind
-            && self.key_kind == request.key_kind
-            && self.index.key_space() == request.key_space
-    }
-}
-
-/// A non-empty collection of unfulfillable requests
-#[derive(Clone, PartialEq, Eq, Hash, Debug)]
-pub struct UnfulfillableRequests {
-    /// A non-empty collection of unfulfillable requests
-    unfulfillable: Vec<UnfulfillableRequest>, // we want `Set` but `IndexSet` is not `Hash`
-}
-impl UnfulfillableRequests {
-    /// # Panics
-    /// Panics if `unfulfillable` is empty.
-    pub fn new(unfulfillable: IndexSet<UnfulfillableRequest>) -> Self {
-        assert!(!unfulfillable.is_empty(), "non_empty must not be empty");
-        Self {
-            unfulfillable: unfulfillable.into_iter().collect(),
-        }
-    }
-    pub fn unfulfillable(&self) -> IndexSet<UnfulfillableRequest> {
-        self.unfulfillable.clone().into_iter().collect()
-    }
-
-    pub fn requests(&self) -> IndexSet<DerivationRequest> {
-        self.unfulfillable
-            .clone()
-            .into_iter()
-            .map(|ur| ur.request)
-            .collect()
-    }
-}
-
-/// The outcome of peeking the next derivation index for a request.
-#[derive(Clone, PartialEq, Eq, Debug)]
-pub enum NextDerivationPeekOutcome {
-    /// We failed to peek the next derivation index for the request, probably
-    /// an error while reading from cache.
-    Failure(CommonError),
-
-    /// All requests would have at least one factor left after they are fulfilled.
-    Fulfillable,
-
-    /// The `IndexMap` contains the last consumed index for each request.
-    ///
-    /// N.B. that if some request would not consume the last factor, it will not
-    /// be present in this map.
-    Unfulfillable(UnfulfillableRequests),
-}
-
 /// A cache for pre-derived keys, saved on file and which will derive more keys
 /// if needed, using UI/UX via KeysCollector.
 ///
