+++ src/factor_instance_provider/profile_extensions.rs
@@ -53,24 +53,3 @@ impl Profile {
 }
 
 pub const DERIVATION_INDEX_BATCH_SIZE: HDPathValue = 50;
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct UncoveredEntity {
-    pub on_chain: OnChainEntityState,
-    pub key_hash_to_factor_instances:
-        HashMap<PublicKeyHash, HierarchicalDeterministicFactorInstance>,
-}
-impl UncoveredEntity {
-    pub fn new(
-        on_chain: OnChainEntityState,
-        key_hash_to_factor_instances: HashMap<
-            PublicKeyHash,
-            HierarchicalDeterministicFactorInstance,
-        >,
-    ) -> Self {
-        Self {
-            on_chain,
-            key_hash_to_factor_instances,
-        }
-    }
-}
