+++ src/recovery/mod.rs
@@ -1,7 +1,9 @@
 mod entity_recovery_outcome;
+mod uncovered_entity;
 
 #[allow(clippy::module_inception)]
 mod recovery;
 
 pub use entity_recovery_outcome::*;
 pub use recovery::*;
+pub use uncovered_entity::*;
